- [x] depends on https://github.com/distribution/distribution/pull/3883

Rewrite the regular expressions to use named capturing groups.
This simplifies handling the resulting matches, but does require
some additional handling to associate matches with their names.

Also making some changes to the matching to match how domains
are _actually_ matched; some of that was already handled in
code parsing the results of the regex, but now this is handled
by the regex itself.

Before:

    BenchmarkParse
    BenchmarkParse-10              12696             93805 ns/op            9311 B/op        185 allocs/op
    PASS


After:

    BenchmarkParse
    BenchmarkParse-10              12486             94774 ns/op           18617 B/op        178 allocs/op
    PASS

Benchstat:

    go test -run='^$' -bench=. -count=10 ./reference/ > old.txt
    go test -run='^$' -bench=. -count=10 ./reference/ > new.txt

    benchstat old.txt new.txt
    name       old time/op    new time/op    delta
    Parse-10   91.7µs ± 0%    97.0µs ±11%   +5.82%  (p=0.000 n=9+10)

    name       old alloc/op   new alloc/op   delta
    Parse-10    9.32kB ± 0%   18.63kB ± 0%  +99.93%  (p=0.000 n=10+10)

    name       old allocs/op  new allocs/op  delta
    Parse-10        185 ± 0%       178 ± 0%   -3.78%  (p=0.000 n=10+10)