@thaJeztah 2022-11-25
Looks like ghcr.io is unhappy (tried locally and also unable to pull);

```
Pull down action image 'ghcr.io/opencontainers/distribution-spec/conformance:v1.0.0@sha256:6302582c3b35169f7846282ad5b81471cd8814f30f7a958030e04fd82b19e17f'
  /usr/bin/docker pull ghcr.io/opencontainers/distribution-spec/conformance:v1.0.0@sha256:6302582c3b35169f7846282ad5b81471cd8814f30f7a958030e04fd82b19e17f
  Error response from daemon: Get "https://ghcr.io/v2/opencontainers/distribution-spec/conformance/manifests/sha256:6302582c3b35169f7846282ad5b81471cd8814f30f7a958030e04fd82b19e17f": Get "https://ghcr.io/token?scope=repository%3Aopencontainers%2Fdistribution-spec%2Fconformance%3Apull&service=ghcr.io": context deadline exceeded (Client.Timeout exceeded while awaiting headers)
  Warning: Docker pull failed with exit code 1, back off 7.185 seconds before retry.
  /usr/bin/docker pull ghcr.io/opencontainers/distribution-spec/conformance:v1.0.0@sha256:6302582c3b35169f7846282ad5b81471cd8814f30f7a958030e04fd82b19e17f
  Error response from daemon: Get "https://ghcr.io/v2/opencontainers/distribution-spec/conformance/manifests/sha256:6302582c3b35169f7846282ad5b81471cd8814f30f7a958030e04fd82b19e17f": Get "https://ghcr.io/token?scope=repository%3Aopencontainers%2Fdistribution-spec%2Fconformance%3Apull&service=ghcr.io": net/http: request canceled (Client.Timeout exceeded while awaiting headers)
  Warning: Docker pull failed with exit code 1, back off 8.553 seconds before retry.
  /usr/bin/docker pull ghcr.io/opencontainers/distribution-spec/conformance:v1.0.0@sha256:6302582c3b35169f7846282ad5b81471cd8814f30f7a958030e04fd82b19e17f
  Error response from daemon: Get "https://ghcr.io/v2/opencontainers/distribution-spec/conformance/manifests/sha256:6302582c3b35169f7846282ad5b81471cd8814f30f7a958030e04fd82b19e17f": Get "https://ghcr.io/token?scope=repository%3Aopencontainers%2Fdistribution-spec%2Fconformance%3Apull&service=ghcr.io": net/http: request canceled (Client.Timeout exceeded while awaiting headers)
Error: Docker pull failed with exit code 1
```