@thaJeztah 2023-05-11
@corhere @squizzi @milosgajdos interested to hear your take on this; I think this makes the code/handling easier to reason with, and the Regexes more "complete", and potentially more usable. But (as commented) _does_ come with an increase memory use. Wondering if we think the pros outweigh the cons.