@milosgajdos 2023-12-16
@thaJeztah do you mind moving this PR to the https://github.com/distribution/reference repo?