@milosgajdos 2024-07-04
Closing as `reference` has moved to https://github.com/distribution/reference

We need to re-open the PR there.